+++ src/FluentSerializer.Core/Naming/NamingStrategies/AbstractSpanNamingStrategy.cs
@@ -25,11 +25,20 @@ public abstract class AbstractSpanNamingStrategy : INamingStrategy
 	{
 		Span<char> characterSpan = stackalloc char[name.Length];
 
+#if NETSTANDARD2_0
+		ConvertCasing(name.AsSpan(), ref characterSpan);
+#else
 		ConvertCasing(name, ref characterSpan);
+#endif
 
 		var newName = characterSpan.ToString();
 		Guard.Against.InvalidName(newName);
+
+#if NETSTANDARD2_0
+		return newName.AsSpan();
+#else
 		return newName;
+#endif
 	}
 
 	/// <summary>