+++ src/FluentSerializer.Xml/DataNodes/Nodes/XmlFragment.Writing.cs
@@ -1,8 +1,11 @@
 using Ardalis.GuardClauses;
 
+using FluentSerializer.Core.DataNodes;
 using FluentSerializer.Core.Extensions;
 using FluentSerializer.Core.Text;
 using FluentSerializer.Xml.Configuration;
+
+using Microsoft.Extensions.ObjectPool;
 
 using System.Linq;
 
@@ -12,6 +15,10 @@ public readonly partial struct XmlFragment
 {
 	/// <inheritdoc />
 	public override string ToString() => this.ToString(XmlSerializerConfiguration.Default);
+
+	/// <inheritdoc />
+	public string WriteTo(in ObjectPool<ITextWriter> stringBuilders, in bool format = true, in bool writeNull = true, in int indent = 0) =>
+		DataNodeExtensions.WriteTo(this, in stringBuilders, in format, in writeNull, in indent);
 
 	/// <inheritdoc />
 	public ITextWriter AppendTo(ref ITextWriter stringBuilder, in bool format = true, in int indent = 0, in bool writeNull = true)