+++ src/FluentSerializer.Core/Changelog.md
@@ -2,7 +2,7 @@
 
 ## @next
 
--
+- [#8](https://github.com/Marvin-Brouwer/FluentSerializer/issues/8) Added `netframework` backwards compatibility
 
 ## v3.0.1